+++ server/client/src/pages/Org/CreateOrgForm.jsx
@@ -40,10 +40,10 @@ export default function CreateOrgForm() {
         </FormControl>
         <Stack spacing={6}>
           <Button
-            bg={'blue.400'}
+            bg={'teal.400'}
             color={'white'}
             _hover={{
-              bg: 'blue.500',
+              bg: 'teal.500',
             }}>
             Create Org
           </Button>
